OptPath Object

A R6Class to represent the OptPath.

OptPath

Format

R6Class object

Member Variables

par.set

[ParamSet]
The [ParamSet] from which the x values will be added to this OptPath.

y.names

[character()]
The names for the y values. Default is “y”.

minimize

[logical()]
A logical vector indicating which y components are to be minimized. Per default all are TRUE.

check.feasible

[logical(1)]
Should new x values be checked for feasibility according to the ParamSet.

data

[data.table]
This field contains all values logged into the opt.path.

Methods

add()

[function]
Add a new line to the OptPath.

Active Bindings

x.names

[character()]
Names of the x-values

length

[integer(1)]
The number of entries in this OptPath.

x

[data.table]
A subset of data only containing the x values.

y

[data.table]
A subset of data only containing the y values.

dim

[integer(1)]
The dimensionality of the optimization problem.