Visitor to traverse ParamHandle

A R6Class to visit attached ParamHandle object(host). Using a separate visitor to a class is a frequently used design pattern which could seperate the operations(traversal the graph structure of the ParamHandle) on the ParamHandle and the ParamHandle pointer themselves. The ParamHandle node has two kinds of children node: the mandatory node and conditional node. Mandatory nodes are unconditional children for a graph-like hyper parameter, for example, the normal SVM has a hyperparameter C which is mandatory node but the hyperparameter gamma only makes sense when the hyperparameter kernel type is rbf kernel.

ParamVisitor

Format

R6Class object

Value

[ParamVisitor].

See also

Other ParamHelpers: ParamBase, ParamHandle, ParamNode, ParamSetTree